Crate tauri_nspanel

Re-exports

pub use builder::CollectionBehavior;

pub use builder::PanelBuilder;

pub use builder::PanelLevel;

pub use builder::StyleMask;

pub use builder::TrackingAreaOptions;

Modules

builder

common

event

panel

Macros

panel

Macro to create a custom NSPanel class

panel_event

Macro to create a custom event handler for panels

tauri_panel

Top-level macro that wraps panel and event handler declarations

Structs

AnyObject

An Objective-C object.

NSNotification

************** Notifications ***************

NSObject

The root class of most Objective-C class hierarchies.

NSPanel

Apple’s documentation

NSResponder

Apple’s documentation

NSView

Apple’s documentation

NSWindow

Apple’s documentation

Store

WebviewPanelManager

Enums

Error

Traits

EventHandler

Trait for event handlers that can be used with panels

FromWindow

Trait for panels that can be created from a window

ManagerExt

Panel

Common trait for all panel types

WebviewWindowExt

Functions

init

Initializes the plugin.

Type Aliases

NSPoint

A point in a Cartesian coordinate system.

NSRect

A rectangle.

NSSize

A two-dimensional size.

PanelHandle

Type alias for shared panel references