tauri_nspanel

Struct NSObject

pub struct NSObject { /* private fields */ }

The root class of most Objective-C class hierarchies.

This represents the NSObject class. The name “NSObject” also refers to a protocol, see NSObjectProtocol for that.

This class has been defined in objc since macOS 10.8, but is also re-exported under objc2_foundation::NSObject, you might want to use that path instead.

Implementations

impl NSObject

pub fn new() -> Retained<NSObject>

Create a new empty NSObject.

This method is a shorthand for calling alloc and then init.

pub fn init(this: Allocated<NSObject>) -> Retained<NSObject>

Initialize an already allocated object.

See Apple’s documentation for details.

Example

use objc2::runtime::NSObject;
use objc2::AnyThread;

let obj = NSObject::init(NSObject::alloc());

pub fn doesNotRecognizeSelector(&self, sel: Sel) -> !

Handle messages the object doesn’t recognize.

See Apple’s documentation for details.