tauri_nspanel

Struct NSResponder

#[repr(C)]
pub struct NSResponder { /* private fields */ }

Apple’s documentation

Implementations

impl NSResponder

NSInterfaceStyle.

pub fn interfaceStyle(&self) -> usize

👎Deprecated

pub fn setInterfaceStyle(&self, interface_style: usize)

👎Deprecated

impl NSResponder

pub fn init(this: Allocated<NSResponder>) -> Retained<NSResponder>

pub unsafe fn initWithCoder( this: Allocated<NSResponder>, coder: &NSCoder, ) -> Option<Retained<NSResponder>>

Safety

coder possibly has further requirements.

pub unsafe fn nextResponder(&self) -> Option<Retained<NSResponder>>

Safety

This is not retained internally, you must ensure the object is still alive.

pub unsafe fn setNextResponder(&self, next_responder: Option<&NSResponder>)

Setter for nextResponder.

Safety

This is unretained, you must ensure the object is kept alive while in use.

pub unsafe fn tryToPerform_with( &self, action: Sel, object: Option<&AnyObject>, ) -> bool

Safety

action must be a valid selector.

object should be of the correct type.

pub fn performKeyEquivalent(&self, event: &NSEvent) -> bool

pub fn validRequestorForSendType_returnType( &self, send_type: Option<&NSString>, return_type: Option<&NSString>, ) -> Option<Retained<AnyObject>>

pub fn mouseDown(&self, event: &NSEvent)

pub fn rightMouseDown(&self, event: &NSEvent)

pub fn otherMouseDown(&self, event: &NSEvent)

pub fn mouseUp(&self, event: &NSEvent)

pub fn rightMouseUp(&self, event: &NSEvent)

pub fn otherMouseUp(&self, event: &NSEvent)

pub fn mouseMoved(&self, event: &NSEvent)

pub fn mouseDragged(&self, event: &NSEvent)

pub fn mouseCancelled(&self, event: &NSEvent)

pub fn scrollWheel(&self, event: &NSEvent)

pub fn rightMouseDragged(&self, event: &NSEvent)

pub fn otherMouseDragged(&self, event: &NSEvent)

pub fn mouseEntered(&self, event: &NSEvent)

pub fn mouseExited(&self, event: &NSEvent)

pub fn keyDown(&self, event: &NSEvent)

pub fn keyUp(&self, event: &NSEvent)

pub fn flagsChanged(&self, event: &NSEvent)

pub fn tabletPoint(&self, event: &NSEvent)

pub fn tabletProximity(&self, event: &NSEvent)

pub fn cursorUpdate(&self, event: &NSEvent)

pub fn magnifyWithEvent(&self, event: &NSEvent)

pub fn rotateWithEvent(&self, event: &NSEvent)

pub fn swipeWithEvent(&self, event: &NSEvent)

pub fn beginGestureWithEvent(&self, event: &NSEvent)

pub fn endGestureWithEvent(&self, event: &NSEvent)

pub fn smartMagnifyWithEvent(&self, event: &NSEvent)

pub fn changeModeWithEvent(&self, event: &NSEvent)

pub fn touchesBeganWithEvent(&self, event: &NSEvent)

pub fn touchesMovedWithEvent(&self, event: &NSEvent)

pub fn touchesEndedWithEvent(&self, event: &NSEvent)

pub fn touchesCancelledWithEvent(&self, event: &NSEvent)

pub fn quickLookWithEvent(&self, event: &NSEvent)

pub fn pressureChangeWithEvent(&self, event: &NSEvent)

pub fn contextMenuKeyDown(&self, event: &NSEvent)

Handle a key event that should present a context menu at the user focus.

Most applications should not override this method. Instead, you should customize the context menu displayed from a keyboard event by implementing menuForEvent: and selectionAnchorRect, or showContextMenuForSelection:, rather than this method.

You should only override this method when you do not want the system-provided default behavior for the context menu hotkey, either for a specific key combination, or for the hotkey in general. For example, if your application already provides a different behavior for control-Return (the default context menu hotkey definition), and you want to preserve that behavior, you should override this method to handle that specific key combination, and then return without calling super. Note that the user may customize the hotkey to a different key combination, so in this example, if any other key combination is passed to your method, you would call super.

An implementation of this method should call [super contextMenuKeyDown:event] to pass the request up the responder chain. If the message reaches the application object, NSApplication’s implementation of this method will send showContextMenuForSelection: to the responder chain. If you do not call super, then no further handling of the key event will be performed.

Note: In some cases, showContextMenuForSelection: will be called without a prior call to contextMenuKeyDown:. This occurs when a view receives an Accessibility ShowMenu action, or when the user has created a Cocoa Text key binding to map a different key combination to the showContextMenuForSelection: action.

Parameter event: The key down event that matches the system-wide context menu hotkey combination.

See also: showContextMenuForSelection:

pub unsafe fn noResponderFor(&self, event_selector: Sel)

Safety

event_selector must be a valid selector.

pub fn acceptsFirstResponder(&self) -> bool

pub fn becomeFirstResponder(&self) -> bool

pub fn resignFirstResponder(&self) -> bool

pub fn interpretKeyEvents(&self, event_array: &NSArray<NSEvent>)

pub fn flushBufferedKeyEvents(&self)

pub fn menu(&self) -> Option<Retained<NSMenu>>

pub unsafe fn setMenu(&self, menu: Option<&NSMenu>)

Setter for menu.

pub unsafe fn showContextHelp(&self, sender: Option<&AnyObject>)

Safety

sender should be of the correct type.

pub fn helpRequested(&self, event_ptr: &NSEvent)

pub fn shouldBeTreatedAsInkEvent(&self, event: &NSEvent) -> bool

pub fn wantsScrollEventsForSwipeTrackingOnAxis( &self, axis: NSEventGestureAxis, ) -> bool

pub fn wantsForwardedScrollEventsForAxis( &self, axis: NSEventGestureAxis, ) -> bool

pub unsafe fn supplementalTargetForAction_sender( &self, action: Sel, sender: Option<&AnyObject>, ) -> Option<Retained<AnyObject>>

Safety

action must be a valid selector.

sender should be of the correct type.

impl NSResponder

Methods declared on superclass NSObject.

pub fn new(mtm: MainThreadMarker) -> Retained<NSResponder>

impl NSResponder

This impl block contains no items.

NSStandardKeyBindingMethods.

impl NSResponder

NSUndoSupport.

pub fn undoManager(&self) -> Option<Retained<NSUndoManager>>

impl NSResponder

NSControlEditingSupport.

pub fn validateProposedFirstResponder_forEvent( &self, responder: &NSResponder, event: Option<&NSEvent>, ) -> bool

impl NSResponder

NSErrorPresentation.

pub unsafe fn presentError_modalForWindow_delegate_didPresentSelector_contextInfo( &self, error: &NSError, window: &NSWindow, delegate: Option<&AnyObject>, did_present_selector: Option<Sel>, context_info: *mut c_void, )

Safety

delegate should be of the correct type.

did_present_selector must be a valid selector.

context_info must be a valid pointer or null.

pub fn presentError(&self, error: &NSError) -> bool

pub fn willPresentError(&self, error: &NSError) -> Retained<NSError>

impl NSResponder

NSTextFinderSupport.

pub unsafe fn performTextFinderAction(&self, sender: Option<&AnyObject>)

Safety

sender should be of the correct type.

impl NSResponder

NSWindowTabbing.

pub unsafe fn newWindowForTab(&self, sender: Option<&AnyObject>)

Safety

sender should be of the correct type.

impl NSResponder

NSWritingToolsSupport.

pub unsafe fn showWritingTools(&self, sender: Option<&AnyObject>)

Safety

sender should be of the correct type.

impl NSResponder

NSDeprecated.

pub fn performMnemonic(&self, string: &NSString) -> bool

👎Deprecated: This has always returned NO and had no effect on macOS

impl NSResponder

NSTouchBarProvider.

pub fn touchBar(&self) -> Option<Retained<NSTouchBar>>

pub fn setTouchBar(&self, touch_bar: Option<&NSTouchBar>)

Setter for touchBar.

pub fn makeTouchBar(&self) -> Option<Retained<NSTouchBar>>

impl NSResponder

NSUserActivity.

pub fn userActivity(&self) -> Option<Retained<NSUserActivity>>

pub fn setUserActivity(&self, user_activity: Option<&NSUserActivity>)

Setter for userActivity.

pub fn updateUserActivityState(&self, user_activity: &NSUserActivity)

impl NSResponder

NSRestorableState.

pub unsafe fn encodeRestorableStateWithCoder(&self, coder: &NSCoder)

Safety

coder possibly has further requirements.

pub unsafe fn encodeRestorableStateWithCoder_backgroundQueue( &self, coder: &NSCoder, queue: &NSOperationQueue, )

Safety

coder possibly has further requirements.

queue possibly has additional threading requirements.

pub unsafe fn restoreStateWithCoder(&self, coder: &NSCoder)

Safety

coder possibly has further requirements.

pub fn invalidateRestorableState(&self)

pub fn restorableStateKeyPaths( mtm: MainThreadMarker, ) -> Retained<NSArray<NSString>>

pub fn allowedClassesForRestorableStateKeyPath( key_path: &NSString, mtm: MainThreadMarker, ) -> Retained<NSArray<AnyClass>>

When secure state restoration is used, values at restorableStateKeyPaths must support NSSecureCoding and this method will be consulted when restoring state. Any values not of an allowed class will not be set.