tauri_nspanel

Struct NSView

pub struct NSView { /* private fields */ }

Apple’s documentation

Implementations

impl NSView

NSCandidateListTouchBarItem.

pub fn candidateListTouchBarItem( &self, ) -> Option<Retained<NSCandidateListTouchBarItem>>

impl NSView

NSClipViewSuperview.

pub fn reflectScrolledClipView(&self, clip_view: &NSClipView)

pub fn scrollClipView_toPoint(&self, clip_view: &NSClipView, point: CGPoint)

impl NSView

NSConstraintBasedLayoutInstallingConstraints.

pub fn leadingAnchor(&self) -> Retained<NSLayoutXAxisAnchor>

pub fn trailingAnchor(&self) -> Retained<NSLayoutXAxisAnchor>

pub fn leftAnchor(&self) -> Retained<NSLayoutXAxisAnchor>

pub fn rightAnchor(&self) -> Retained<NSLayoutXAxisAnchor>

pub fn topAnchor(&self) -> Retained<NSLayoutYAxisAnchor>

pub fn bottomAnchor(&self) -> Retained<NSLayoutYAxisAnchor>

pub fn widthAnchor(&self) -> Retained<NSLayoutDimension>

pub fn heightAnchor(&self) -> Retained<NSLayoutDimension>

pub fn centerXAnchor(&self) -> Retained<NSLayoutXAxisAnchor>

pub fn centerYAnchor(&self) -> Retained<NSLayoutYAxisAnchor>

pub fn firstBaselineAnchor(&self) -> Retained<NSLayoutYAxisAnchor>

pub fn lastBaselineAnchor(&self) -> Retained<NSLayoutYAxisAnchor>

pub fn constraints(&self) -> Retained<NSArray<NSLayoutConstraint>>

pub fn addConstraint(&self, constraint: &NSLayoutConstraint)

pub fn addConstraints(&self, constraints: &NSArray<NSLayoutConstraint>)

pub fn removeConstraint(&self, constraint: &NSLayoutConstraint)

pub fn removeConstraints(&self, constraints: &NSArray<NSLayoutConstraint>)

impl NSView

NSConstraintBasedLayoutCoreMethods.

pub fn updateConstraintsForSubtreeIfNeeded(&self)

pub fn updateConstraints(&self)

pub fn needsUpdateConstraints(&self) -> bool

pub fn setNeedsUpdateConstraints(&self, needs_update_constraints: bool)

Setter for needsUpdateConstraints.

impl NSView

NSConstraintBasedCompatibility.

pub fn translatesAutoresizingMaskIntoConstraints(&self) -> bool

pub fn setTranslatesAutoresizingMaskIntoConstraints( &self, translates_autoresizing_mask_into_constraints: bool, )

Setter for translatesAutoresizingMaskIntoConstraints.

pub fn requiresConstraintBasedLayout(mtm: MainThreadMarker) -> bool

impl NSView

NSConstraintBasedLayoutLayering.

pub fn alignmentRectForFrame(&self, frame: CGRect) -> CGRect

pub fn frameForAlignmentRect(&self, alignment_rect: CGRect) -> CGRect

pub fn alignmentRectInsets(&self) -> NSEdgeInsets

pub fn firstBaselineOffsetFromTop(&self) -> f64

pub fn lastBaselineOffsetFromBottom(&self) -> f64

pub fn baselineOffsetFromBottom(&self) -> f64

pub fn intrinsicContentSize(&self) -> CGSize

pub fn invalidateIntrinsicContentSize(&self)

pub fn contentHuggingPriorityForOrientation( &self, orientation: NSLayoutConstraintOrientation, ) -> f32

pub fn setContentHuggingPriority_forOrientation( &self, priority: f32, orientation: NSLayoutConstraintOrientation, )

pub fn contentCompressionResistancePriorityForOrientation( &self, orientation: NSLayoutConstraintOrientation, ) -> f32

pub fn setContentCompressionResistancePriority_forOrientation( &self, priority: f32, orientation: NSLayoutConstraintOrientation, )

pub fn isHorizontalContentSizeConstraintActive(&self) -> bool

pub fn setHorizontalContentSizeConstraintActive( &self, horizontal_content_size_constraint_active: bool, )

Setter for isHorizontalContentSizeConstraintActive.

pub fn isVerticalContentSizeConstraintActive(&self) -> bool

pub fn setVerticalContentSizeConstraintActive( &self, vertical_content_size_constraint_active: bool, )

Setter for isVerticalContentSizeConstraintActive.

impl NSView

NSConstraintBasedLayoutFittingSize.

pub fn fittingSize(&self) -> CGSize

impl NSView

NSConstraintBasedLayoutDebugging.

pub fn constraintsAffectingLayoutForOrientation( &self, orientation: NSLayoutConstraintOrientation, ) -> Retained<NSArray<NSLayoutConstraint>>

pub fn hasAmbiguousLayout(&self) -> bool

pub fn exerciseAmbiguityInLayout(&self)

impl NSView

NSLayoutGuideSupport.

pub fn addLayoutGuide(&self, guide: &NSLayoutGuide)

pub fn removeLayoutGuide(&self, guide: &NSLayoutGuide)

pub fn layoutGuides(&self) -> Retained<NSArray<NSLayoutGuide>>

impl NSView

NSViewEnclosingMenuItem.

pub fn enclosingMenuItem(&self) -> Option<Retained<NSMenuItem>>

impl NSView

NSOpenGLSurfaceResolution.

pub fn wantsBestResolutionOpenGLSurface(&self) -> bool

👎Deprecated: Use NSOpenGLView instead.

pub fn setWantsBestResolutionOpenGLSurface( &self, wants_best_resolution_open_gl_surface: bool, )

👎Deprecated: Use NSOpenGLView instead.

Setter for wantsBestResolutionOpenGLSurface.

impl NSView

NSExtendedDynamicRange.

pub fn wantsExtendedDynamicRangeOpenGLSurface(&self) -> bool

👎Deprecated: Use NSOpenGLView instead.

pub fn setWantsExtendedDynamicRangeOpenGLSurface( &self, wants_extended_dynamic_range_open_gl_surface: bool, )

👎Deprecated: Use NSOpenGLView instead.

Setter for wantsExtendedDynamicRangeOpenGLSurface.

impl NSView

NSPressureConfiguration.

pub fn pressureConfiguration(&self) -> Option<Retained<NSPressureConfiguration>>

pub fn setPressureConfiguration( &self, pressure_configuration: Option<&NSPressureConfiguration>, )

Setter for pressureConfiguration.

impl NSView

NSRulerMarkerClientViewDelegation.

pub fn rulerView_shouldMoveMarker( &self, ruler: &NSRulerView, marker: &NSRulerMarker, ) -> bool

pub fn rulerView_willMoveMarker_toLocation( &self, ruler: &NSRulerView, marker: &NSRulerMarker, location: f64, ) -> f64

pub fn rulerView_didMoveMarker( &self, ruler: &NSRulerView, marker: &NSRulerMarker, )

pub fn rulerView_shouldRemoveMarker( &self, ruler: &NSRulerView, marker: &NSRulerMarker, ) -> bool

pub fn rulerView_didRemoveMarker( &self, ruler: &NSRulerView, marker: &NSRulerMarker, )

pub fn rulerView_shouldAddMarker( &self, ruler: &NSRulerView, marker: &NSRulerMarker, ) -> bool

pub fn rulerView_willAddMarker_atLocation( &self, ruler: &NSRulerView, marker: &NSRulerMarker, location: f64, ) -> f64

pub fn rulerView_didAddMarker( &self, ruler: &NSRulerView, marker: &NSRulerMarker, )

pub fn rulerView_handleMouseDown(&self, ruler: &NSRulerView, event: &NSEvent)

pub fn rulerView_willSetClientView( &self, ruler: &NSRulerView, new_client: &NSView, )

pub fn rulerView_locationForPoint( &self, ruler: &NSRulerView, point: CGPoint, ) -> f64

pub fn rulerView_pointForLocation( &self, ruler: &NSRulerView, point: f64, ) -> CGPoint

impl NSView

pub fn initWithFrame( this: Allocated<NSView>, frame_rect: CGRect, ) -> Retained<NSView>

pub unsafe fn initWithCoder( this: Allocated<NSView>, coder: &NSCoder, ) -> Option<Retained<NSView>>

Safety

coder possibly has further requirements.

pub fn window(&self) -> Option<Retained<NSWindow>>

pub unsafe fn superview(&self) -> Option<Retained<NSView>>

Safety

This is not retained internally, you must ensure the object is still alive.

pub fn subviews(&self) -> Retained<NSArray<NSView>>

pub fn setSubviews(&self, subviews: &NSArray<NSView>)

Setter for subviews.

This is copied when set.

pub fn isDescendantOf(&self, view: &NSView) -> bool

pub fn ancestorSharedWithView(&self, view: &NSView) -> Option<Retained<NSView>>

pub unsafe fn opaqueAncestor(&self) -> Option<Retained<NSView>>

Safety

This is not retained internally, you must ensure the object is still alive.

pub fn isHidden(&self) -> bool

pub fn setHidden(&self, hidden: bool)

Setter for isHidden.

pub fn isHiddenOrHasHiddenAncestor(&self) -> bool

pub unsafe fn getRectsBeingDrawn_count( &self, rects: *mut *const CGRect, count: *mut isize, )

Safety

rects must be a valid pointer or null.

count must be a valid pointer or null.

pub fn needsToDrawRect(&self, rect: CGRect) -> bool

pub fn wantsDefaultClipping(&self) -> bool

pub fn viewDidHide(&self)

pub fn viewDidUnhide(&self)

pub fn addSubview(&self, view: &NSView)

pub fn addSubview_positioned_relativeTo( &self, view: &NSView, place: NSWindowOrderingMode, other_view: Option<&NSView>, )

pub unsafe fn sortSubviewsUsingFunction_context( &self, compare: unsafe extern "C-unwind" fn(NonNull<NSView>, NonNull<NSView>, *mut c_void) -> NSComparisonResult, context: *mut c_void, )

Safety

compare must be implemented correctly.

context must be a valid pointer or null.

pub fn viewWillMoveToWindow(&self, new_window: Option<&NSWindow>)

pub fn viewDidMoveToWindow(&self)

pub fn viewWillMoveToSuperview(&self, new_superview: Option<&NSView>)

pub fn viewDidMoveToSuperview(&self)

pub fn didAddSubview(&self, subview: &NSView)

pub fn willRemoveSubview(&self, subview: &NSView)

pub fn removeFromSuperview(&self)

pub fn replaceSubview_with(&self, old_view: &NSView, new_view: &NSView)

pub fn removeFromSuperviewWithoutNeedingDisplay(&self)

pub fn viewDidChangeBackingProperties(&self)

pub fn postsFrameChangedNotifications(&self) -> bool

pub fn setPostsFrameChangedNotifications( &self, posts_frame_changed_notifications: bool, )

Setter for postsFrameChangedNotifications.

pub fn resizeSubviewsWithOldSize(&self, old_size: CGSize)

pub fn resizeWithOldSuperviewSize(&self, old_size: CGSize)

pub fn autoresizesSubviews(&self) -> bool

pub fn setAutoresizesSubviews(&self, autoresizes_subviews: bool)

Setter for autoresizesSubviews.

pub fn autoresizingMask(&self) -> NSAutoresizingMaskOptions

pub fn setAutoresizingMask(&self, autoresizing_mask: NSAutoresizingMaskOptions)

Setter for autoresizingMask.

pub fn setFrameOrigin(&self, new_origin: CGPoint)

pub fn setFrameSize(&self, new_size: CGSize)

pub fn frame(&self) -> CGRect

pub fn setFrame(&self, frame: CGRect)

Setter for frame.

pub fn frameRotation(&self) -> f64

pub fn setFrameRotation(&self, frame_rotation: f64)

Setter for frameRotation.

pub fn frameCenterRotation(&self) -> f64

pub fn setFrameCenterRotation(&self, frame_center_rotation: f64)

Setter for frameCenterRotation.

pub fn setBoundsOrigin(&self, new_origin: CGPoint)

pub fn setBoundsSize(&self, new_size: CGSize)

pub fn boundsRotation(&self) -> f64

pub fn setBoundsRotation(&self, bounds_rotation: f64)

Setter for boundsRotation.

pub fn translateOriginToPoint(&self, translation: CGPoint)

pub fn scaleUnitSquareToSize(&self, new_unit_size: CGSize)

pub fn rotateByAngle(&self, angle: f64)

pub fn bounds(&self) -> CGRect

pub fn setBounds(&self, bounds: CGRect)

Setter for bounds.

pub fn isFlipped(&self) -> bool

pub fn isRotatedFromBase(&self) -> bool

pub fn isRotatedOrScaledFromBase(&self) -> bool

pub fn isOpaque(&self) -> bool

pub fn convertPoint_fromView( &self, point: CGPoint, view: Option<&NSView>, ) -> CGPoint

pub fn convertPoint_toView( &self, point: CGPoint, view: Option<&NSView>, ) -> CGPoint

pub fn convertSize_fromView( &self, size: CGSize, view: Option<&NSView>, ) -> CGSize

pub fn convertSize_toView(&self, size: CGSize, view: Option<&NSView>) -> CGSize

pub fn convertRect_fromView( &self, rect: CGRect, view: Option<&NSView>, ) -> CGRect

pub fn convertRect_toView(&self, rect: CGRect, view: Option<&NSView>) -> CGRect

pub fn backingAlignedRect_options( &self, rect: CGRect, options: NSAlignmentOptions, ) -> CGRect

pub fn centerScanRect(&self, rect: CGRect) -> CGRect

pub fn convertPointToBacking(&self, point: CGPoint) -> CGPoint

pub fn convertPointFromBacking(&self, point: CGPoint) -> CGPoint

pub fn convertSizeToBacking(&self, size: CGSize) -> CGSize

pub fn convertSizeFromBacking(&self, size: CGSize) -> CGSize

pub fn convertRectToBacking(&self, rect: CGRect) -> CGRect

pub fn convertRectFromBacking(&self, rect: CGRect) -> CGRect

pub fn convertPointToLayer(&self, point: CGPoint) -> CGPoint

pub fn convertPointFromLayer(&self, point: CGPoint) -> CGPoint

pub fn convertSizeToLayer(&self, size: CGSize) -> CGSize

pub fn convertSizeFromLayer(&self, size: CGSize) -> CGSize

pub fn convertRectToLayer(&self, rect: CGRect) -> CGRect

pub fn convertRectFromLayer(&self, rect: CGRect) -> CGRect

pub fn canDrawConcurrently(&self) -> bool

pub unsafe fn setCanDrawConcurrently(&self, can_draw_concurrently: bool)

Setter for canDrawConcurrently.

pub fn canDraw(&self) -> bool

👎Deprecated: If a view needs display, -drawRect: or -updateLayer will be called automatically when the view is able to draw. To check whether a view is in a window, call -window. To check whether a view is hidden, call -isHiddenOrHasHiddenAncestor.

pub fn setNeedsDisplayInRect(&self, invalid_rect: CGRect)

pub fn needsDisplay(&self) -> bool

pub fn setNeedsDisplay(&self, needs_display: bool)

Setter for needsDisplay.

pub fn lockFocus(&self)

👎Deprecated: To draw, subclass NSView and implement -drawRect:; AppKit’s automatic deferred display mechanism will call -drawRect: as necessary to display the view.

pub fn unlockFocus(&self)

👎Deprecated: To draw, subclass NSView and implement -drawRect:; AppKit’s automatic deferred display mechanism will call -drawRect: as necessary to display the view.

pub fn lockFocusIfCanDraw(&self) -> bool

👎Deprecated: To draw, subclass NSView and implement -drawRect:; AppKit’s automatic deferred display mechanism will call -drawRect: as necessary to display the view.

pub fn lockFocusIfCanDrawInContext(&self, context: &NSGraphicsContext) -> bool

👎Deprecated: Use -[NSView displayRectIgnoringOpacity:inContext:] to draw a view subtree into a graphics context.

pub fn focusView(mtm: MainThreadMarker) -> Option<Retained<NSView>>

pub fn visibleRect(&self) -> CGRect

The portion of the view that isn’t clipped by its superviews.

Visibility, as reflected by this property, doesn’t account for whether other view or window objects overlap the current view or whether the current view is installed in a window at all. This value of this property is NSZeroRect if the current view is effectively hidden.

During a printing operation, the visible rectangle is further clipped to the page being imaged.

pub fn display(&self)

pub fn displayIfNeeded(&self)

pub fn displayIfNeededIgnoringOpacity(&self)

pub fn displayRect(&self, rect: CGRect)

pub fn displayIfNeededInRect(&self, rect: CGRect)

pub fn displayRectIgnoringOpacity(&self, rect: CGRect)

pub fn displayIfNeededInRectIgnoringOpacity(&self, rect: CGRect)

pub fn drawRect(&self, dirty_rect: CGRect)

pub fn displayRectIgnoringOpacity_inContext( &self, rect: CGRect, context: &NSGraphicsContext, )

pub fn bitmapImageRepForCachingDisplayInRect( &self, rect: CGRect, ) -> Option<Retained<NSBitmapImageRep>>

pub fn cacheDisplayInRect_toBitmapImageRep( &self, rect: CGRect, bitmap_image_rep: &NSBitmapImageRep, )

pub fn viewWillDraw(&self)

pub fn scrollPoint(&self, point: CGPoint)

pub fn scrollRectToVisible(&self, rect: CGRect) -> bool

pub fn autoscroll(&self, event: &NSEvent) -> bool

pub fn adjustScroll(&self, new_visible: CGRect) -> CGRect

pub fn scrollRect_by(&self, rect: CGRect, delta: CGSize)

👎Deprecated: Use NSScrollView to achieve scrolling views.

pub fn translateRectsNeedingDisplayInRect_by( &self, clip_rect: CGRect, delta: CGSize, )

pub fn hitTest(&self, point: CGPoint) -> Option<Retained<NSView>>

pub fn mouse_inRect(&self, point: CGPoint, rect: CGRect) -> bool

pub fn viewWithTag(&self, tag: isize) -> Option<Retained<NSView>>

pub fn tag(&self) -> isize

pub fn performKeyEquivalent(&self, event: &NSEvent) -> bool

pub fn acceptsFirstMouse(&self, event: Option<&NSEvent>) -> bool

pub fn shouldDelayWindowOrderingForEvent(&self, event: &NSEvent) -> bool

pub fn needsPanelToBecomeKey(&self) -> bool

pub fn mouseDownCanMoveWindow(&self) -> bool

pub fn acceptsTouchEvents(&self) -> bool

👎Deprecated: Use allowedTouchTypes instead

pub fn setAcceptsTouchEvents(&self, accepts_touch_events: bool)

👎Deprecated: Use allowedTouchTypes instead

Setter for acceptsTouchEvents.

pub fn wantsRestingTouches(&self) -> bool

pub fn setWantsRestingTouches(&self, wants_resting_touches: bool)

Setter for wantsRestingTouches.

pub fn makeBackingLayer(&self) -> Retained<CALayer>

pub fn layerContentsRedrawPolicy(&self) -> NSViewLayerContentsRedrawPolicy

pub fn setLayerContentsRedrawPolicy( &self, layer_contents_redraw_policy: NSViewLayerContentsRedrawPolicy, )

Setter for layerContentsRedrawPolicy.

pub fn layerContentsPlacement(&self) -> NSViewLayerContentsPlacement

pub fn setLayerContentsPlacement( &self, layer_contents_placement: NSViewLayerContentsPlacement, )

Setter for layerContentsPlacement.

pub fn wantsLayer(&self) -> bool

pub fn setWantsLayer(&self, wants_layer: bool)

Setter for wantsLayer.

pub fn layer(&self) -> Option<Retained<CALayer>>

pub fn setLayer(&self, layer: Option<&CALayer>)

Setter for layer.

pub fn wantsUpdateLayer(&self) -> bool

pub fn updateLayer(&self)

pub fn canDrawSubviewsIntoLayer(&self) -> bool

pub fn setCanDrawSubviewsIntoLayer(&self, can_draw_subviews_into_layer: bool)

Setter for canDrawSubviewsIntoLayer.

pub fn layoutSubtreeIfNeeded(&self)

pub fn layout(&self)

pub fn needsLayout(&self) -> bool

pub fn setNeedsLayout(&self, needs_layout: bool)

Setter for needsLayout.

pub fn alphaValue(&self) -> f64

pub fn setAlphaValue(&self, alpha_value: f64)

Setter for alphaValue.

pub fn layerUsesCoreImageFilters(&self) -> bool

pub fn setLayerUsesCoreImageFilters(&self, layer_uses_core_image_filters: bool)

Setter for layerUsesCoreImageFilters.

pub fn backgroundFilters(&self) -> Retained<NSArray<CIFilter>>

pub fn setBackgroundFilters(&self, background_filters: &NSArray<CIFilter>)

Setter for backgroundFilters.

This is copied when set.

pub fn compositingFilter(&self) -> Option<Retained<CIFilter>>

pub fn setCompositingFilter(&self, compositing_filter: Option<&CIFilter>)

Setter for compositingFilter.

pub fn contentFilters(&self) -> Retained<NSArray<CIFilter>>

pub fn setContentFilters(&self, content_filters: &NSArray<CIFilter>)

Setter for contentFilters.

This is copied when set.

pub fn shadow(&self) -> Option<Retained<NSShadow>>

pub fn setShadow(&self, shadow: Option<&NSShadow>)

Setter for shadow.

This is copied when set.

pub fn clipsToBounds(&self) -> bool

pub fn setClipsToBounds(&self, clips_to_bounds: bool)

Setter for clipsToBounds.

pub fn postsBoundsChangedNotifications(&self) -> bool

pub fn setPostsBoundsChangedNotifications( &self, posts_bounds_changed_notifications: bool, )

Setter for postsBoundsChangedNotifications.

pub fn enclosingScrollView(&self) -> Option<Retained<NSScrollView>>

pub fn menuForEvent(&self, event: &NSEvent) -> Option<Retained<NSMenu>>

pub fn defaultMenu(mtm: MainThreadMarker) -> Option<Retained<NSMenu>>

pub fn willOpenMenu_withEvent(&self, menu: &NSMenu, event: &NSEvent)

A contextual menu is being opened from the receiving view. The view should update any visual state in response — such as making a selection.

Parameter menu: The contextual menu that is being opened on the view

Parameter event: The event that caused the menu to open.

pub fn didCloseMenu_withEvent(&self, menu: &NSMenu, event: Option<&NSEvent>)

A contextual menu shown from the receiving view has been closed. This is only called if the menu had been opened and the view previously received -willOpenMenu:withEvent:.The view should update any visual state in response — such as removing a temporary selection.

Parameter menu: The contextual menu that was open on the view

Parameter event: The event that caused the menu to close. This may be nil if there is no specific event that triggered the closing.

pub fn toolTip(&self) -> Option<Retained<NSString>>

pub fn setToolTip(&self, tool_tip: Option<&NSString>)

Setter for toolTip.

This is copied when set.

pub unsafe fn addToolTipRect_owner_userData( &self, rect: CGRect, owner: &AnyObject, data: *mut c_void, ) -> isize

Safety

owner should be of the correct type.

data must be a valid pointer or null.

pub fn removeToolTip(&self, tag: isize)

pub fn removeAllToolTips(&self)

pub fn viewWillStartLiveResize(&self)

pub fn viewDidEndLiveResize(&self)

pub fn inLiveResize(&self) -> bool

pub fn preservesContentDuringLiveResize(&self) -> bool

pub fn rectPreservedDuringLiveResize(&self) -> CGRect

pub fn inputContext(&self) -> Option<Retained<NSTextInputContext>>

pub fn rectForSmartMagnificationAtPoint_inRect( &self, location: CGPoint, visible_rect: CGRect, ) -> CGRect

pub fn userInterfaceLayoutDirection(&self) -> NSUserInterfaceLayoutDirection

pub fn setUserInterfaceLayoutDirection( &self, user_interface_layout_direction: NSUserInterfaceLayoutDirection, )

Setter for userInterfaceLayoutDirection.

pub fn prepareForReuse(&self)

pub fn isCompatibleWithResponsiveScrolling(mtm: MainThreadMarker) -> bool

pub fn prepareContentInRect(&self, rect: CGRect)

pub fn preparedContentRect(&self) -> CGRect

pub fn setPreparedContentRect(&self, prepared_content_rect: CGRect)

Setter for preparedContentRect.

pub fn allowsVibrancy(&self) -> bool

pub fn viewDidChangeEffectiveAppearance(&self)

Override point for reacting to the effective appearance of the receiver changing. At this point effectiveAppearance property reflects the new appearance.

impl NSView

Methods declared on superclass NSResponder.

pub fn init(this: Allocated<NSView>) -> Retained<NSView>

impl NSView

Methods declared on superclass NSObject.

pub fn new(mtm: MainThreadMarker) -> Retained<NSView>

impl NSView

NSKeyboardUI.

pub unsafe fn nextKeyView(&self) -> Option<Retained<NSView>>

Safety

This is not retained internally, you must ensure the object is still alive.

pub unsafe fn setNextKeyView(&self, next_key_view: Option<&NSView>)

Setter for nextKeyView.

Safety

This is unretained, you must ensure the object is kept alive while in use.

pub unsafe fn previousKeyView(&self) -> Option<Retained<NSView>>

Safety

This is not retained internally, you must ensure the object is still alive.

pub unsafe fn nextValidKeyView(&self) -> Option<Retained<NSView>>

Safety

This is not retained internally, you must ensure the object is still alive.

pub unsafe fn previousValidKeyView(&self) -> Option<Retained<NSView>>

Safety

This is not retained internally, you must ensure the object is still alive.

pub fn canBecomeKeyView(&self) -> bool

pub fn setKeyboardFocusRingNeedsDisplayInRect(&self, rect: CGRect)

pub fn focusRingType(&self) -> NSFocusRingType

pub fn setFocusRingType(&self, focus_ring_type: NSFocusRingType)

Setter for focusRingType.

pub fn defaultFocusRingType(mtm: MainThreadMarker) -> NSFocusRingType

pub fn drawFocusRingMask(&self)

pub fn focusRingMaskBounds(&self) -> CGRect

pub fn noteFocusRingMaskChanged(&self)

impl NSView

NSPrinting.

pub fn writeEPSInsideRect_toPasteboard( &self, rect: CGRect, pasteboard: &NSPasteboard, )

pub fn dataWithEPSInsideRect(&self, rect: CGRect) -> Retained<NSData>

pub fn writePDFInsideRect_toPasteboard( &self, rect: CGRect, pasteboard: &NSPasteboard, )

pub fn dataWithPDFInsideRect(&self, rect: CGRect) -> Retained<NSData>

pub unsafe fn print(&self, sender: Option<&AnyObject>)

Safety

sender should be of the correct type.

pub unsafe fn knowsPageRange(&self, range: *mut NSRange) -> bool

Safety

range must be a valid pointer.

pub fn heightAdjustLimit(&self) -> f64

pub fn widthAdjustLimit(&self) -> f64

pub unsafe fn adjustPageWidthNew_left_right_limit( &self, new_right: NonNull<f64>, old_left: f64, old_right: f64, right_limit: f64, )

Safety

new_right must be a valid pointer.

pub unsafe fn adjustPageHeightNew_top_bottom_limit( &self, new_bottom: NonNull<f64>, old_top: f64, old_bottom: f64, bottom_limit: f64, )

Safety

new_bottom must be a valid pointer.

pub fn rectForPage(&self, page: isize) -> CGRect

pub fn locationOfPrintRect(&self, rect: CGRect) -> CGPoint

pub fn drawPageBorderWithSize(&self, border_size: CGSize)

pub fn pageHeader(&self) -> Retained<NSAttributedString>

pub fn pageFooter(&self) -> Retained<NSAttributedString>

pub fn drawSheetBorderWithSize(&self, border_size: CGSize)

👎Deprecated: This is never invoked and the NSView implementation does nothing

This method is obsolete. It will never be invoked from within AppKit, and NSView’s implementation of it does nothing. **

pub fn printJobTitle(&self) -> Retained<NSString>

pub fn beginDocument(&self)

pub fn endDocument(&self)

pub fn beginPageInRect_atPlacement(&self, rect: CGRect, location: CGPoint)

pub fn endPage(&self)

impl NSView

NSDrag.

pub fn beginDraggingSessionWithItems_event_source( &self, items: &NSArray<NSDraggingItem>, event: &NSEvent, source: &ProtocolObject<dyn NSDraggingSource>, ) -> Retained<NSDraggingSession>

pub fn registeredDraggedTypes(&self) -> Retained<NSArray<NSString>>

pub fn registerForDraggedTypes(&self, new_types: &NSArray<NSString>)

pub fn unregisterDraggedTypes(&self)

impl NSView

NSFullScreenMode.

pub unsafe fn enterFullScreenMode_withOptions( &self, screen: &NSScreen, options: Option<&NSDictionary<NSString>>, ) -> bool

Safety

options generic should be of the correct type.

pub unsafe fn exitFullScreenModeWithOptions( &self, options: Option<&NSDictionary<NSString>>, )

Safety

options generic should be of the correct type.

pub fn isInFullScreenMode(&self) -> bool

impl NSView

NSDefinition.

pub fn showDefinitionForAttributedString_atPoint( &self, attr_string: Option<&NSAttributedString>, text_baseline_origin: CGPoint, )

pub unsafe fn showDefinitionForAttributedString_range_options_baselineOriginProvider( &self, attr_string: Option<&NSAttributedString>, target_range: NSRange, options: Option<&NSDictionary<NSString>>, origin_provider: Option<&Block<dyn Fn(NSRange) -> CGPoint>>, )

Safety

options generic should be of the correct type.

impl NSView

NSFindIndicator.

pub fn isDrawingFindIndicator(&self) -> bool

impl NSView

NSGestureRecognizer.

pub fn gestureRecognizers(&self) -> Retained<NSArray<NSGestureRecognizer>>

pub fn setGestureRecognizers( &self, gesture_recognizers: &NSArray<NSGestureRecognizer>, )

Setter for gestureRecognizers.

This is copied when set.

pub fn addGestureRecognizer(&self, gesture_recognizer: &NSGestureRecognizer)

pub fn removeGestureRecognizer(&self, gesture_recognizer: &NSGestureRecognizer)

impl NSView

NSTouchBar.

pub fn allowedTouchTypes(&self) -> NSTouchTypeMask

pub fn setAllowedTouchTypes(&self, allowed_touch_types: NSTouchTypeMask)

Setter for allowedTouchTypes.

impl NSView

NSSafeAreas.

pub fn safeAreaInsets(&self) -> NSEdgeInsets

pub fn additionalSafeAreaInsets(&self) -> NSEdgeInsets

pub fn setAdditionalSafeAreaInsets( &self, additional_safe_area_insets: NSEdgeInsets, )

Setter for additionalSafeAreaInsets.

pub fn safeAreaLayoutGuide(&self) -> Retained<NSLayoutGuide>

pub fn safeAreaRect(&self) -> CGRect

pub fn layoutMarginsGuide(&self) -> Retained<NSLayoutGuide>

impl NSView

NSCompactControlSizeMetrics.

pub fn prefersCompactControlSizeMetrics(&self) -> bool

When this property is true, any NSControls in the view or its descendants will be sized with compact metrics compatible with macOS 15 and earlier. Defaults to false

pub fn setPrefersCompactControlSizeMetrics( &self, prefers_compact_control_size_metrics: bool, )

Setter for prefersCompactControlSizeMetrics.

impl NSView

NSTrackingArea.

pub fn addTrackingArea(&self, tracking_area: &NSTrackingArea)

pub fn removeTrackingArea(&self, tracking_area: &NSTrackingArea)

pub fn trackingAreas(&self) -> Retained<NSArray<NSTrackingArea>>

pub fn updateTrackingAreas(&self)

pub fn addCursorRect_cursor(&self, rect: CGRect, object: &NSCursor)

pub fn removeCursorRect_cursor(&self, rect: CGRect, object: &NSCursor)

pub fn discardCursorRects(&self)

pub fn resetCursorRects(&self)

pub unsafe fn addTrackingRect_owner_userData_assumeInside( &self, rect: CGRect, owner: &AnyObject, data: *mut c_void, flag: bool, ) -> isize

Safety

owner should be of the correct type.

data must be a valid pointer or null.

pub fn removeTrackingRect(&self, tag: isize)

impl NSView

NSDisplayLink.

pub unsafe fn displayLinkWithTarget_selector( &self, target: &AnyObject, selector: Sel, ) -> Retained<CADisplayLink>

Safety

target should be of the correct type.

selector must be a valid selector.

impl NSView

NSDeprecated.

pub unsafe fn dragImage_at_offset_event_pasteboard_source_slideBack( &self, image: &NSImage, view_location: CGPoint, initial_offset: CGSize, event: &NSEvent, pboard: &NSPasteboard, source_obj: &AnyObject, slide_flag: bool, )

👎Deprecated: Use -beginDraggingSessionWithItems:event:source: instead

Safety

source_obj should be of the correct type.

pub fn dragFile_fromRect_slideBack_event( &self, filename: &NSString, rect: CGRect, flag: bool, event: &NSEvent, ) -> bool

👎Deprecated: Use -beginDraggingSessionWithItems:event:source: instead

pub unsafe fn dragPromisedFilesOfTypes_fromRect_source_slideBack_event( &self, type_array: &NSArray<NSString>, rect: CGRect, source_object: &AnyObject, flag: bool, event: &NSEvent, ) -> bool

👎Deprecated: Use -beginDraggingSessionWithItems:event:source: with an NSFilePromiseProvider instead

Safety

source_object should be of the correct type.

pub fn convertPointToBase(&self, point: CGPoint) -> CGPoint

👎Deprecated

pub fn convertPointFromBase(&self, point: CGPoint) -> CGPoint

👎Deprecated

pub fn convertSizeToBase(&self, size: CGSize) -> CGSize

👎Deprecated

pub fn convertSizeFromBase(&self, size: CGSize) -> CGSize

👎Deprecated

pub fn convertRectToBase(&self, rect: CGRect) -> CGRect

👎Deprecated

pub fn convertRectFromBase(&self, rect: CGRect) -> CGRect

👎Deprecated

pub fn performMnemonic(&self, string: &NSString) -> bool

👎Deprecated: This has always returned NO and had no effect on macOS

pub fn shouldDrawColor(&self) -> bool

👎Deprecated: This method no longer does anything

pub fn gState(&self) -> isize

👎Deprecated

pub fn allocateGState(&self)

👎Deprecated

pub fn setUpGState(&self)

👎Deprecated

pub fn renewGState(&self)

👎Deprecated

impl NSView

NSWritingToolsCoordinator.

pub fn writingToolsCoordinator( &self, ) -> Option<Retained<NSWritingToolsCoordinator>>

pub fn setWritingToolsCoordinator( &self, writing_tools_coordinator: Option<&NSWritingToolsCoordinator>, )

Setter for writingToolsCoordinator.

impl NSView

LayoutRegions.

pub fn layoutGuideForLayoutRegion( &self, layout_region: &NSViewLayoutRegion, ) -> Retained<NSLayoutGuide>

pub fn edgeInsetsForLayoutRegion( &self, layout_region: &NSViewLayoutRegion, ) -> NSEdgeInsets

pub fn rectForLayoutRegion(&self, layout_region: &NSViewLayoutRegion) -> CGRect