tauri_nspanel

Struct NSWindow

pub struct NSWindow { /* private fields */ }

Apple’s documentation

Implementations

impl NSWindow

NSDrawers.

pub fn drawers(&self) -> Option<Retained<NSArray<NSDrawer>>>

👎Deprecated:
Drawers are deprecated; consider using NSSplitViewController

impl NSWindow

NSConstraintBasedLayoutCoreMethods.

pub fn updateConstraintsIfNeeded(&self)

pub fn layoutIfNeeded(&self)

impl NSWindow

NSConstraintBasedLayoutAnchoring.

pub fn anchorAttributeForOrientation( &self, orientation: NSLayoutConstraintOrientation, ) -> NSLayoutAttribute

pub fn setAnchorAttribute_forOrientation( &self, attr: NSLayoutAttribute, orientation: NSLayoutConstraintOrientation, )

impl NSWindow

NSConstraintBasedLayoutDebugging.

pub fn visualizeConstraints( &self, constraints: Option<&NSArray<NSLayoutConstraint>>, )

impl NSWindow

pub fn frameRectForContentRect_styleMask( c_rect: CGRect, style: NSWindowStyleMask, mtm: MainThreadMarker, ) -> CGRect

pub fn contentRectForFrameRect_styleMask( f_rect: CGRect, style: NSWindowStyleMask, mtm: MainThreadMarker, ) -> CGRect

pub fn minFrameWidthWithTitle_styleMask( title: &NSString, style: NSWindowStyleMask, mtm: MainThreadMarker, ) -> f64

pub fn defaultDepthLimit(mtm: MainThreadMarker) -> NSWindowDepth

pub fn frameRectForContentRect(&self, content_rect: CGRect) -> CGRect

pub fn contentRectForFrameRect(&self, frame_rect: CGRect) -> CGRect

pub unsafe fn initWithContentRect_styleMask_backing_defer( this: Allocated<NSWindow>, content_rect: CGRect, style: NSWindowStyleMask, backing_store_type: NSBackingStoreType, flag: bool, ) -> Retained<NSWindow>

pub unsafe fn initWithContentRect_styleMask_backing_defer_screen( this: Allocated<NSWindow>, content_rect: CGRect, style: NSWindowStyleMask, backing_store_type: NSBackingStoreType, flag: bool, screen: Option<&NSScreen>, ) -> Retained<NSWindow>

pub unsafe fn initWithCoder( this: Allocated<NSWindow>, coder: &NSCoder, ) -> Retained<NSWindow>

Safety

coder possibly has further requirements.

pub fn title(&self) -> Retained<NSString>

pub fn setTitle(&self, title: &NSString)

Setter for title.

This is copied when set.

pub fn subtitle(&self) -> Retained<NSString>

Secondary text that may be displayed adjacent to or below the primary title depending on the configuration of the window. A value of empty string will remove the subtitle from the window layout.

pub fn setSubtitle(&self, subtitle: &NSString)

Setter for subtitle.

This is copied when set.

pub fn titleVisibility(&self) -> NSWindowTitleVisibility

See the enum values for how this property works.

pub fn setTitleVisibility(&self, title_visibility: NSWindowTitleVisibility)

Setter for titleVisibility.

pub fn titlebarAppearsTransparent(&self) -> bool

When YES,the titlebar doesn’t draw its background, allowing all buttons to show through, and “click through” to happen. In general, this is only useful when NSFullSizeContentViewWindowMaskis set.

pub fn setTitlebarAppearsTransparent(&self, titlebar_appears_transparent: bool)

Setter for titlebarAppearsTransparent.

pub fn toolbarStyle(&self) -> NSWindowToolbarStyle

Specifies how the titlebar area of the window should appear when the window displays an NSToolbar

pub fn setToolbarStyle(&self, toolbar_style: NSWindowToolbarStyle)

Setter for toolbarStyle.

pub fn contentLayoutRect(&self) -> CGRect

The contentLayoutRectwill return the area inside the window that is for non-obscured content. Typically, this is the same thing as the contentView’s frame. However, for windows with the NSFullSizeContentViewWindowMaskset, there needs to be a way to determine the portion that is not under the toolbar. The contentLayoutRectreturns the portion of the layout that is not obscured under the toolbar. contentLayoutRectis in window coordinates. It is KVO compliant. */

pub fn contentLayoutGuide(&self) -> Option<Retained<AnyObject>>

contentLayoutGuideis a corollary to contentLayoutRect.It can be used by autolayout constraints to automatically bind to the contentLayoutRect.

pub fn titlebarAccessoryViewControllers( &self, ) -> Retained<NSArray<NSTitlebarAccessoryViewController>>

pub fn setTitlebarAccessoryViewControllers( &self, titlebar_accessory_view_controllers: &NSArray<NSTitlebarAccessoryViewController>, )

Setter for titlebarAccessoryViewControllers.

This is copied when set.

pub fn addTitlebarAccessoryViewController( &self, child_view_controller: &NSTitlebarAccessoryViewController, )

pub fn insertTitlebarAccessoryViewController_atIndex( &self, child_view_controller: &NSTitlebarAccessoryViewController, index: isize, )

pub fn removeTitlebarAccessoryViewControllerAtIndex(&self, index: isize)

pub fn representedURL(&self) -> Option<Retained<NSURL>>

If url is not nil and its path is not empty, the window will show a document icon in the titlebar. If the url represents a filename or other resource with a known icon, that icon will be used as the document icon. Otherwise the default document icon will be used. The icon can be customized using -[[NSWindow standardWindowButton:NSWindowDocumentIconButton] setImage:customImage]. If url is not nil and its path is not empty, the window will have a pop-up menu which can be shown via command-click on the area containing the document icon and title. By default, this menu will display the path components of the url. The presence and contents of this menu can be controlled by the delegate method -[window:shouldPopUpDocumentPathMenu:] If the url is nil or has an empty path, the window will not show a document icon and will not have a pop-up menu available via command-click.

pub fn setRepresentedURL(&self, represented_url: Option<&NSURL>)

Setter for representedURL.

This is copied when set.

pub fn representedFilename(&self) -> Retained<NSString>

pub fn setRepresentedFilename(&self, represented_filename: &NSString)

Setter for representedFilename.

This is copied when set.

pub fn setTitleWithRepresentedFilename(&self, filename: &NSString)

pub fn isExcludedFromWindowsMenu(&self) -> bool

pub fn setExcludedFromWindowsMenu(&self, excluded_from_windows_menu: bool)

Setter for isExcludedFromWindowsMenu.

pub fn contentView(&self) -> Option<Retained<NSView>>

pub fn setContentView(&self, content_view: Option<&NSView>)

Setter for contentView.

pub fn delegate(&self) -> Option<Retained<ProtocolObject<dyn NSWindowDelegate>>>

pub fn setDelegate( &self, delegate: Option<&ProtocolObject<dyn NSWindowDelegate>>, )

Setter for delegate.

This is a [weak property][objc2::topics::weak_property].

pub fn windowNumber(&self) -> isize

pub fn styleMask(&self) -> NSWindowStyleMask

Note: The styleMask can only be set on macOS 10.6 and later. Valid styleMasksettings have the same restrictions as the styleMaskpassed to -initWithContentRect:styleMask:backing:defer:. Some styleMaskchanges will cause the view hierarchy to be rebuilt, since there is a different subclass for the top level view of a borderless window than for the top level view of a titled window.

pub fn setStyleMask(&self, style_mask: NSWindowStyleMask)

Setter for styleMask.

pub unsafe fn fieldEditor_forObject( &self, create_flag: bool, object: Option<&AnyObject>, ) -> Option<Retained<NSText>>

Safety

object should be of the correct type.

pub unsafe fn endEditingFor(&self, object: Option<&AnyObject>)

Safety

object should be of the correct type.

pub fn constrainFrameRect_toScreen( &self, frame_rect: CGRect, screen: Option<&NSScreen>, ) -> CGRect

pub fn setFrame_display(&self, frame_rect: CGRect, flag: bool)

pub fn setContentSize(&self, size: CGSize)

pub fn setFrameOrigin(&self, point: CGPoint)

pub fn setFrameTopLeftPoint(&self, point: CGPoint)

pub fn cascadeTopLeftFromPoint(&self, top_left_point: CGPoint) -> CGPoint

pub fn cascadingReferenceFrame(&self) -> CGRect

The frame to use when cascading or sizing a new window based on the receiver’s position or size. This may be different from frame when the receiver is positioned by the system.

pub fn frame(&self) -> CGRect

pub fn animationResizeTime(&self, new_frame: CGRect) -> f64

Subclasses can override animationResizeTime:to control the total time for the frame change. newFrameis the rect passed into setFrame:display:animate:

pub fn setFrame_display_animate( &self, frame_rect: CGRect, display_flag: bool, animate_flag: bool, )

setFrame:display:animate:is equivalent to setFrame:display:if the animateFlagis NO. If the animationFlagis YES, this method will perform a smooth resize of the window, where the total time for the resize is specified by -animationResizeTime:

pub fn inLiveResize(&self) -> bool

pub fn resizeIncrements(&self) -> CGSize

pub fn setResizeIncrements(&self, resize_increments: CGSize)

Setter for resizeIncrements.

pub fn aspectRatio(&self) -> CGSize

pub fn setAspectRatio(&self, aspect_ratio: CGSize)

Setter for aspectRatio.

pub fn contentResizeIncrements(&self) -> CGSize

pub fn setContentResizeIncrements(&self, content_resize_increments: CGSize)

Setter for contentResizeIncrements.

pub fn contentAspectRatio(&self) -> CGSize

pub fn setContentAspectRatio(&self, content_aspect_ratio: CGSize)

Setter for contentAspectRatio.

pub fn viewsNeedDisplay(&self) -> bool

pub fn setViewsNeedDisplay(&self, views_need_display: bool)

Setter for viewsNeedDisplay.

pub fn displayIfNeeded(&self)

pub fn display(&self)

pub fn preservesContentDuringLiveResize(&self) -> bool

pub fn setPreservesContentDuringLiveResize( &self, preserves_content_during_live_resize: bool, )

Setter for preservesContentDuringLiveResize.

pub fn update(&self)

pub fn makeFirstResponder(&self, responder: Option<&NSResponder>) -> bool

pub fn firstResponder(&self) -> Option<Retained<NSResponder>>

pub fn resizeFlags(&self) -> NSEventModifierFlags

pub fn close(&self)

pub fn isReleasedWhenClosed(&self) -> bool

pub unsafe fn setReleasedWhenClosed(&self, released_when_closed: bool)

Setter for isReleasedWhenClosed.

pub fn miniaturize(&self, sender: Option<&AnyObject>)

pub fn deminiaturize(&self, sender: Option<&AnyObject>)

pub fn isZoomed(&self) -> bool

pub fn zoom(&self, sender: Option<&AnyObject>)

pub fn isMiniaturized(&self) -> bool

pub unsafe fn tryToPerform_with( &self, action: Sel, object: Option<&AnyObject>, ) -> bool

Safety

action must be a valid selector.

object should be of the correct type.

pub fn validRequestorForSendType_returnType( &self, send_type: Option<&NSString>, return_type: Option<&NSString>, ) -> Option<Retained<AnyObject>>

pub fn backgroundColor(&self) -> Retained<NSColor>

pub fn setBackgroundColor(&self, background_color: Option<&NSColor>)

Setter for backgroundColor.

This is copied when set.

pub fn setContentBorderThickness_forEdge( &self, thickness: f64, edge: NSRectEdge, )

pub fn contentBorderThicknessForEdge(&self, edge: NSRectEdge) -> f64

pub fn setAutorecalculatesContentBorderThickness_forEdge( &self, flag: bool, edge: NSRectEdge, )

pub fn autorecalculatesContentBorderThicknessForEdge( &self, edge: NSRectEdge, ) -> bool

pub fn isMovable(&self) -> bool

pub fn setMovable(&self, movable: bool)

Setter for isMovable.

pub fn isMovableByWindowBackground(&self) -> bool

pub fn setMovableByWindowBackground(&self, movable_by_window_background: bool)

Setter for isMovableByWindowBackground.

pub fn hidesOnDeactivate(&self) -> bool

pub fn setHidesOnDeactivate(&self, hides_on_deactivate: bool)

Setter for hidesOnDeactivate.

pub fn canHide(&self) -> bool

Indicates whether a window can be hidden during -[NSApplication hide:]. Default is YES.

pub fn setCanHide(&self, can_hide: bool)

Setter for canHide.

pub fn center(&self)

pub fn makeKeyAndOrderFront(&self, sender: Option<&AnyObject>)

pub fn orderFront(&self, sender: Option<&AnyObject>)

pub fn orderBack(&self, sender: Option<&AnyObject>)

pub fn orderOut(&self, sender: Option<&AnyObject>)

pub fn orderWindow_relativeTo( &self, place: NSWindowOrderingMode, other_win: isize, )

pub fn orderFrontRegardless(&self)

pub fn miniwindowImage(&self) -> Option<Retained<NSImage>>

pub fn setMiniwindowImage(&self, miniwindow_image: Option<&NSImage>)

Setter for miniwindowImage.

pub fn miniwindowTitle(&self) -> Retained<NSString>

pub fn setMiniwindowTitle(&self, miniwindow_title: Option<&NSString>)

Setter for miniwindowTitle.

This is copied when set.

pub fn dockTile(&self) -> Retained<NSDockTile>

pub fn isDocumentEdited(&self) -> bool

pub fn setDocumentEdited(&self, document_edited: bool)

Setter for isDocumentEdited.

pub fn isVisible(&self) -> bool

pub fn isKeyWindow(&self) -> bool

pub fn isMainWindow(&self) -> bool

pub fn canBecomeKeyWindow(&self) -> bool

pub fn canBecomeMainWindow(&self) -> bool

pub fn makeKeyWindow(&self)

Makes the window key and main if eligible, updating NSApplication’s -keyWindow and -mainWindow properties.

pub fn makeMainWindow(&self)

Makes the window main if eligible. Updates NSApplication’s -mainWindow property.

pub fn becomeKeyWindow(&self)

Informs the window that it has become the key window. This method exists as an override point. Do not invoke directly. Instead, invoke -makeKeyWindow.

pub fn resignKeyWindow(&self)

Informs the window that it has stopped being the key window. This method exists as an override point. Do not invoke directly. Windows automatically receive this message when deactivating or when another window has become key.

pub fn becomeMainWindow(&self)

Informs the window that it has become the main window. This method exists as an override point. Do not invoke directly. Instead, invoke -makeMainWindow.

pub fn resignMainWindow(&self)

Informs the window that it has stopped being the main window. This method exists as an override point. Do not invoke directly. Windows automatically receive this message when deactivating or when another window has become main.

pub fn worksWhenModal(&self) -> bool

pub fn preventsApplicationTerminationWhenModal(&self) -> bool

A Boolean value that indicates whether or not to prevent application termination when the receiving window is presented modally. The value of this property is YES if the window should prevent application termination when modal; otherwise, NO. The default value is YES. However, note that some window subclasses and some windows created indirectly (like those created by UI frameworks like AppKit and SwiftUI), may have different default values. For example, the Open panel and toolbar customization sheets should not prevent application termination, so those windows have preventsApplicationTerminationWhenModal set to NO. Some NSAlerts, like those that are simply informational, have windows that do not prevent application termination by default. Setting this property overrides the default behavior.

pub fn setPreventsApplicationTerminationWhenModal( &self, prevents_application_termination_when_modal: bool, )

Setter for preventsApplicationTerminationWhenModal.

pub fn convertRectToScreen(&self, rect: CGRect) -> CGRect

pub fn convertRectFromScreen(&self, rect: CGRect) -> CGRect

pub fn convertPointToScreen(&self, point: CGPoint) -> CGPoint

pub fn convertPointFromScreen(&self, point: CGPoint) -> CGPoint

pub fn convertRectToBacking(&self, rect: CGRect) -> CGRect

pub fn convertRectFromBacking(&self, rect: CGRect) -> CGRect

pub fn convertPointToBacking(&self, point: CGPoint) -> CGPoint

pub fn convertPointFromBacking(&self, point: CGPoint) -> CGPoint

pub fn backingAlignedRect_options( &self, rect: CGRect, options: NSAlignmentOptions, ) -> CGRect

Use NSIntegralRectWithOptions() to produce a backing store pixel aligned rectangle from the given input rectangle in window coordinates.

pub fn backingScaleFactor(&self) -> f64

Returns the scale factor representing the number of backing store pixels corresponding to each linear unit in window space on this NSWindow.This method is provided for rare cases when the explicit scale factor is needed. Please use -convert*ToBacking: methods whenever possible.

pub fn performClose(&self, sender: Option<&AnyObject>)

pub fn performMiniaturize(&self, sender: Option<&AnyObject>)

pub fn performZoom(&self, sender: Option<&AnyObject>)

pub fn dataWithEPSInsideRect(&self, rect: CGRect) -> Retained<NSData>

pub fn dataWithPDFInsideRect(&self, rect: CGRect) -> Retained<NSData>

pub fn print(&self, sender: Option<&AnyObject>)

pub fn allowsToolTipsWhenApplicationIsInactive(&self) -> bool

Default is NO.Set to YESto allow a window to display tooltips even when the application is in the background. Note that, enabling tooltips in an inactive application will cause the app to do work any time the mouse passes over the window. This can degrade system performance. Returns YESif this window displays tooltips even when the application is in the background. To configure this setting you should call -setAllowsToolTipsWhenApplicationIsInactive: instead of overriding -allowsToolTipsWhenApplicationIsInactive.

pub fn setAllowsToolTipsWhenApplicationIsInactive( &self, allows_tool_tips_when_application_is_inactive: bool, )

Setter for allowsToolTipsWhenApplicationIsInactive.

pub fn backingType(&self) -> NSBackingStoreType

pub fn setBackingType(&self, backing_type: NSBackingStoreType)

Setter for backingType.

pub fn level(&self) -> isize

pub fn setLevel(&self, level: isize)

Setter for level.

pub fn depthLimit(&self) -> NSWindowDepth

pub fn setDepthLimit(&self, depth_limit: NSWindowDepth)

Setter for depthLimit.

pub fn setDynamicDepthLimit(&self, flag: bool)

pub fn hasDynamicDepthLimit(&self) -> bool

pub fn screen(&self) -> Option<Retained<NSScreen>>

The screen property returns the best screen for the window. If the window only intersects one screen, it returns that screen. If it intersects more than one screen, then it resolves the tie through based on what space it is mostly on. It may return nil if there are no available screens, or it is completely off screen.

pub fn deepestScreen(&self) -> Option<Retained<NSScreen>>

pub fn hasShadow(&self) -> bool

pub fn setHasShadow(&self, has_shadow: bool)

Setter for hasShadow.

pub fn invalidateShadow(&self)

pub fn alphaValue(&self) -> f64

pub fn setAlphaValue(&self, alpha_value: f64)

Setter for alphaValue.

pub fn isOpaque(&self) -> bool

pub fn setOpaque(&self, opaque: bool)

Setter for isOpaque.

pub fn sharingType(&self) -> NSWindowSharingType

-setSharingType: specifies whether the window content can be read from another process. The default sharing type is NSWindowSharingReadOnly,which means other processes can read the window content (eg. for window capture) but cannot modify it. If you set your window sharing type to NSWindowSharingNone,so that the content cannot be captured, your window will also not be able to participate in a number of system services, so this setting should be used with caution.

pub fn setSharingType(&self, sharing_type: NSWindowSharingType)

Setter for sharingType.

pub fn allowsConcurrentViewDrawing(&self) -> bool

Controls whether threading of view drawing should be enabled for this window. Defaults to YES.When this is set to YES,AppKit’s view system is allowed to perform -drawRect: activity for the window’s views on threads other than the main thread, for views that have canDrawConcurrently == YES. When this is set to NO,the window’s views will be drawn serially as on 10.5 and earlier, even though some of the views may have canDrawConcurrently == YES.

pub unsafe fn setAllowsConcurrentViewDrawing( &self, allows_concurrent_view_drawing: bool, )

Setter for allowsConcurrentViewDrawing.

pub fn displaysWhenScreenProfileChanges(&self) -> bool

pub fn setDisplaysWhenScreenProfileChanges( &self, displays_when_screen_profile_changes: bool, )

Setter for displaysWhenScreenProfileChanges.

pub fn canBecomeVisibleWithoutLogin(&self) -> bool

This API controls whether the receiver is permitted onscreen before the user has logged in. This property is off by default. Alert panels and windows presented by input managers are examples of windows which should have this property set.

pub fn setCanBecomeVisibleWithoutLogin( &self, can_become_visible_without_login: bool, )

Setter for canBecomeVisibleWithoutLogin.

pub fn collectionBehavior(&self) -> NSWindowCollectionBehavior

pub fn setCollectionBehavior( &self, collection_behavior: NSWindowCollectionBehavior, )

Setter for collectionBehavior.

pub fn animationBehavior(&self) -> NSWindowAnimationBehavior

Provides for per-window control over automatic orderFront/orderOut animation behaviors added in 10.7. Can be set to NSWindowAnimationBehaviorNoneto disable Appkit’s automatic animations for a given window, or to one of the other non-Default NSWindowAnimationBehaviorvalues to override AppKit’s automatic inference of appropriate animation behavior based on the window’s apparent type.

pub fn setAnimationBehavior( &self, animation_behavior: NSWindowAnimationBehavior, )

Setter for animationBehavior.

pub fn isOnActiveSpace(&self) -> bool

Returns YESif this window is associated with the active space. For visible windows, this API indicates whether the window is currently visible on the active space. For offscreen windows, it indicates whether ordering the window onscreen would make it bring it onto the active space

pub fn toggleFullScreen(&self, sender: Option<&AnyObject>)

-toggleFullScreen: enters or exits for full screen. A window must have NSWindowCollectionBehaviorFullScreenAuxiliaryor NSWindowCollectionBehaviorFullScreenPrimaryincluded in the collectionBehaviorproperty; if it does not, this method may simply do nothing.

pub fn stringWithSavedFrame(&self) -> Retained<NSString>

pub fn setFrameFromString(&self, string: &NSString)

pub fn saveFrameUsingName(&self, name: &NSString)

pub fn setFrameUsingName_force(&self, name: &NSString, force: bool) -> bool

pub fn setFrameUsingName(&self, name: &NSString) -> bool

pub fn setFrameAutosaveName(&self, name: &NSString) -> bool

pub fn frameAutosaveName(&self) -> Retained<NSString>

pub fn removeFrameUsingName(name: &NSString, mtm: MainThreadMarker)

pub fn minSize(&self) -> CGSize

pub fn setMinSize(&self, min_size: CGSize)

Setter for minSize.

pub fn maxSize(&self) -> CGSize

pub fn setMaxSize(&self, max_size: CGSize)

Setter for maxSize.

pub fn contentMinSize(&self) -> CGSize

pub fn setContentMinSize(&self, content_min_size: CGSize)

Setter for contentMinSize.

pub fn contentMaxSize(&self) -> CGSize

pub fn setContentMaxSize(&self, content_max_size: CGSize)

Setter for contentMaxSize.

pub fn minFullScreenContentSize(&self) -> CGSize

pub fn setMinFullScreenContentSize(&self, min_full_screen_content_size: CGSize)

Setter for minFullScreenContentSize.

pub fn maxFullScreenContentSize(&self) -> CGSize

pub fn setMaxFullScreenContentSize(&self, max_full_screen_content_size: CGSize)

Setter for maxFullScreenContentSize.

pub fn deviceDescription(&self) -> Retained<NSDictionary<NSString>>

pub fn windowController(&self) -> Option<Retained<NSWindowController>>

pub fn setWindowController( &self, window_controller: Option<&NSWindowController>, )

Setter for windowController.

This is a [weak property][objc2::topics::weak_property].

pub fn beginSheet_completionHandler( &self, sheet_window: &NSWindow, handler: Option<&Block<dyn Fn(isize)>>, )

pub fn beginCriticalSheet_completionHandler( &self, sheet_window: &NSWindow, handler: Option<&Block<dyn Fn(isize)>>, )

pub fn endSheet(&self, sheet_window: &NSWindow)

pub fn endSheet_returnCode(&self, sheet_window: &NSWindow, return_code: isize)

pub fn sheets(&self) -> Retained<NSArray<NSWindow>>

An ordered array of the sheets on the window. This consists of the presented sheets in top-to-bottom order, followed by queued sheets in the order they were queued. This does not include nested/sub-sheets.

pub fn attachedSheet(&self) -> Option<Retained<NSWindow>>

Returns the top-most sheet if there is one or more sheets, or nil if there is no sheet.

pub fn isSheet(&self) -> bool

pub fn sheetParent(&self) -> Option<Retained<NSWindow>>

Returns the window that the sheet is directly attached to. This is based on the logical attachment of the sheet, not visual attachment. This relationship exists starting when the sheet is begun (using NSApplication's``-beginSheet:modalForWindow:modalDelegate:didEndSelector:contextInfo: or NSWindow's -beginSheet:completionHandler:), and ending once it is ordered out. Returns nil if the window is not a sheet or has no sheet parent.

pub fn standardWindowButton_forStyleMask( b: NSWindowButton, style_mask: NSWindowStyleMask, mtm: MainThreadMarker, ) -> Option<Retained<NSButton>>

pub fn standardWindowButton( &self, b: NSWindowButton, ) -> Option<Retained<NSButton>>

pub unsafe fn addChildWindow_ordered( &self, child_win: &NSWindow, place: NSWindowOrderingMode, )

pub fn removeChildWindow(&self, child_win: &NSWindow)

pub fn childWindows(&self) -> Option<Retained<NSArray<NSWindow>>>

pub fn parentWindow(&self) -> Option<Retained<NSWindow>>

pub unsafe fn setParentWindow(&self, parent_window: Option<&NSWindow>)

Setter for parentWindow.

This is a [weak property][objc2::topics::weak_property].

pub fn appearanceSource(&self) -> Option<Retained<NSObject>>

If set, the receiver will inherit the appearance of that object, as well as use KVO to observe its effectiveAppearance for changes. Typically this is used for child windows that are shown from a parent window or specific view. Defaults to NSApp.

pub unsafe fn setAppearanceSource(&self, appearance_source: Option<&NSObject>)

Setter for appearanceSource.

This is a [weak property][objc2::topics::weak_property].

Safety

appearance_source must implement NSAppearanceCustomization.

appearance_source might not allow None.

pub fn colorSpace(&self) -> Option<Retained<NSColorSpace>>

pub fn setColorSpace(&self, color_space: Option<&NSColorSpace>)

Setter for colorSpace.

pub fn canRepresentDisplayGamut(&self, display_gamut: NSDisplayGamut) -> bool

-canRepresentDisplayGamut: returns YESif the colorSpace of the receiving window, and the colorSpaceof the screen containing that window, are capable of representing the given display gamut

pub fn windowNumbersWithOptions( options: NSWindowNumberListOptions, mtm: MainThreadMarker, ) -> Option<Retained<NSArray<NSNumber>>>

+windowNumbersWithOptions: returns an autoreleased array of NSNumberscontaining windowNumbers for all visible windows satisfying options. If no options are specified, only visible windows belonging to the calling application and on the active space are included. If options include NSWindowNumberListAllApplications,visible windows belonging to all applications are included. If options include NSWindowNumberListAllSpaces,visible windows on all spaces are included. Windows on the active space are returned in z-order. Examples: To get an array of windowNumbers visible on the current space and belonging to the calling application: windowNumbers = [NSWindow windowNumbersWithOptions:0]; To get an array of windowNumbers visible on any space and belonging to any application: windowNumbers = [NSWindow windowNumbersWithOptions:NSWindowNumberListAllApplications|NSWindowNumberListAllSpaces]; To get an array of windowNumbers visible on any space and belonging to the calling application: windowNumbers = [NSWindow windowNumbersWithOptions:NSWindowNumberListAllSpaces];

pub fn windowNumberAtPoint_belowWindowWithWindowNumber( point: CGPoint, window_number: isize, mtm: MainThreadMarker, ) -> isize

+windowNumberAtPoint:belowWindowWithWindowNumber: returns the number of the frontmost window that would be hit by a mouseDown at the screen location “point”. “windowNum” can be specified to exclude a given window along with all windows above it, and may belong to any application. If no windows are to be excluded, specify 0 for “windowNum”. The windowNumber returned may correspond to a window in another application.

pub fn occlusionState(&self) -> NSWindowOcclusionState

pub fn titlebarSeparatorStyle(&self) -> NSTitlebarSeparatorStyle

Specifies the style of separator displayed between the window’s titlebar and content.

The default value is NSTitlebarSeparatorStyleAutomatic. Changing this value will override any preference made by NSSplitViewItem.

pub fn setTitlebarSeparatorStyle( &self, titlebar_separator_style: NSTitlebarSeparatorStyle, )

Setter for titlebarSeparatorStyle.

pub fn contentViewController(&self) -> Option<Retained<NSViewController>>

The main content view controller for the window. This provides the contentView of the window. Assigning this value will remove the existing contentView and will make the contentViewController.view the main contentView for the window. The default value is nil. The contentViewController only controls the contentView, and not the title of the window. The window title can easily be bound to the contentViewController with the following: [window bind:NSTitleBinding toObject:contentViewController withKeyPath: “ title“ options:nil]. Setting the contentViewController will cause the window to resize based on the current size of the contentViewController. Autolayout should be used to restrict the size of the window. The value of the contentViewController is encoded in the NIB. Directly assigning a contentView will clear out the contentViewController.

pub fn setContentViewController( &self, content_view_controller: Option<&NSViewController>, )

Setter for contentViewController.

pub fn windowWithContentViewController( content_view_controller: &NSViewController, ) -> Retained<NSWindow>

Convenience method for creating an autoreleased titled window with the given contentViewController. A basic NSWindow with the following attributes is made: titled, closable, resizable, miniaturizable. The window’s title is automatically bound to the contentViewController’s title. The size of the window can easily be controlled by utilizing autolayout and applying size constraints to the view (or its subviews). The window has isReleasedWhenClosed set to NO, and it must be explicitly retained to keep the window instance alive. To have it automatically be freed when it is closed, do the following: [window retain] and [window setReleasedWhenClosed:YES].

pub fn performWindowDragWithEvent(&self, event: &NSEvent)

Call to start a drag (moving the window) in the Window Server process. In general, this can be done after a mouseDown event has come in and been examined by an application or view. The view may determine it wants to allow that portion of the window to start a window drag, and can hand off the work to the Window Server process by calling this method. This allows the window to participate in space switching, and other system features. Pass the original mouseDown event to the method. The method will return right away, and a mouseUp may not get sent.

pub fn initialFirstResponder(&self) -> Option<Retained<NSView>>

pub fn setInitialFirstResponder(&self, initial_first_responder: Option<&NSView>)

Setter for initialFirstResponder.

This is a [weak property][objc2::topics::weak_property].

pub fn selectNextKeyView(&self, sender: Option<&AnyObject>)

pub fn selectPreviousKeyView(&self, sender: Option<&AnyObject>)

pub fn selectKeyViewFollowingView(&self, view: &NSView)

pub fn selectKeyViewPrecedingView(&self, view: &NSView)

pub fn keyViewSelectionDirection(&self) -> NSSelectionDirection

pub fn defaultButtonCell(&self) -> Option<Retained<NSButtonCell>>

pub fn setDefaultButtonCell(&self, default_button_cell: Option<&NSButtonCell>)

Setter for defaultButtonCell.

pub fn disableKeyEquivalentForDefaultButtonCell(&self)

pub fn enableKeyEquivalentForDefaultButtonCell(&self)

pub fn autorecalculatesKeyViewLoop(&self) -> bool

pub fn setAutorecalculatesKeyViewLoop( &self, autorecalculates_key_view_loop: bool, )

Setter for autorecalculatesKeyViewLoop.

pub fn recalculateKeyViewLoop(&self)

pub fn toolbar(&self) -> Option<Retained<NSToolbar>>

pub fn setToolbar(&self, toolbar: Option<&NSToolbar>)

Setter for toolbar.

pub fn toggleToolbarShown(&self, sender: Option<&AnyObject>)

pub fn runToolbarCustomizationPalette(&self, sender: Option<&AnyObject>)

pub fn showsToolbarButton(&self) -> bool

👎Deprecated:
This property has no effect

pub fn setShowsToolbarButton(&self, shows_toolbar_button: bool)

👎Deprecated:
This property has no effect

Setter for showsToolbarButton.

pub fn allowsAutomaticWindowTabbing(mtm: MainThreadMarker) -> bool

Allows automatic window tabbing when the value is YES.By default, this will be set to YES,but applications can explicitly opt out of all automatic tabbing by setting it to NO, and can still adopted explicit window tabbing, if desired.

pub fn setAllowsAutomaticWindowTabbing( allows_automatic_window_tabbing: bool, mtm: MainThreadMarker, )

Setter for allowsAutomaticWindowTabbing.

pub fn userTabbingPreference( mtm: MainThreadMarker, ) -> NSWindowUserTabbingPreference

Returns the user’s tabbing preference as set in System Preferences. This value should be queried anytime a new window is made to see if the user wants to automatically show it in tabs.

pub fn tabbingMode(&self) -> NSWindowTabbingMode

Get and set the tabbing mode for this window. This should be set before a window is shown. The default value is NSWindowTabbingModeAutomatic.When the value is NSWindowTabbingModeAutomatic,the system will look at the userTabbingPreferenceand automatically tab windows together based on the tabbingIdentifier, when it is appropriate to do so.

pub fn setTabbingMode(&self, tabbing_mode: NSWindowTabbingMode)

Setter for tabbingMode.

pub fn tabbingIdentifier(&self) -> Retained<NSString>

Windows with the same tabbingIdentifierwill have the ability to be tabbed together when a window is being shown. This allows aggregation of similar windows. By default, the tabbingIdentifierwill be generated based on inherent window properties, such as the window class name, the delegate class name, the window controller class name, and some additional state. Windows can be explicitly made to group together by using the same tabbingIdentifier.

pub fn setTabbingIdentifier(&self, tabbing_identifier: &NSString)

Setter for tabbingIdentifier.

This is copied when set.

pub fn selectNextTab(&self, sender: Option<&AnyObject>)

Actions that can be called to perform various tabbed window behaviors. UI that is hooked up to these items can be automatically validated by calling NSWindow’s validateUserInterfaceItem.

pub fn selectPreviousTab(&self, sender: Option<&AnyObject>)

pub fn moveTabToNewWindow(&self, sender: Option<&AnyObject>)

pub fn mergeAllWindows(&self, sender: Option<&AnyObject>)

pub fn toggleTabBar(&self, sender: Option<&AnyObject>)

pub fn toggleTabOverview(&self, sender: Option<&AnyObject>)

Toggle the Tab Picker / Tab Overview UI which is invoked via “Show All Tabs”. Performs the toggle in an animated fashion. Use tabGroup.isOverviewVisible to find out if it is visible or not at a given time.

pub fn tabbedWindows(&self) -> Option<Retained<NSArray<NSWindow>>>

This is now a cover for self.tabGroup.windows, but will return nil if the window is not showing a tab bar.

pub fn addTabbedWindow_ordered( &self, window: &NSWindow, ordered: NSWindowOrderingMode, )

This is now a cover for -[self.tabGroup addWindow:], which allows more precise placement.

pub fn tab(&self) -> Retained<NSWindowTab>

Access the properties for this window when it is a tabbed window environment. See the NSWindowTabheader and comments for more information.

pub fn tabGroup(&self) -> Option<Retained<NSWindowTabGroup>>

Represents a tab group of windows. This tabGroupis lazily created on demand.

pub fn transferWindowSharingToWindow_completionHandler( &self, window: &NSWindow, completion_handler: &Block<dyn Fn(*mut NSError)>, )

Attempt to move window sharing (i.e. within a SharePlay session) from the receiver to another window. In response to this request, the user may choose to transfer sharing to the new window, or simply stop sharing the content.

Parameter window: A window that is replacing the reciever in representing the user’s current activity.

Parameter completionHandler: A completion block that is called after the request finishes. errorIn the event of a failed transfer request, a non-nil error contains details about the failure.

pub fn hasActiveWindowSharingSession(&self) -> bool

Indicates whether the receiver is the subject of an active SharePlay sharing session.

pub fn requestSharingOfWindow_completionHandler( &self, window: &NSWindow, completion_handler: &Block<dyn Fn(*mut NSError)>, )

Request sharing of window. If there is an available ScreenCaptureKit sharing session, an alert will be presented asking the user to confirm the share

Parameter window: The window to share

Parameter completionHandler: A completion block that is called after the request finishes. errorThe error will be non-nil if the request does not result in a window being shared. The error will be NSUserCancelledError if there is no ScreenCaptureKit session, or if the user rejects the offer to share. If sharing fails for some other reason, the error will provide the details.

pub fn requestSharingOfWindowUsingPreview_title_completionHandler( &self, image: &NSImage, title: &NSString, completion_handler: &Block<dyn Fn(*mut NSError)>, )

Request sharing of window to be provided later. If there is an available ScreenCaptureKit sharing session, an alert will be presented asking the user to confirm the share. The delegate will be asked to provide the window to share via windowForSharingRequestFromWindow:

Parameter image: An image showing a preview of the window to share

Parameter title: The title to show in a confirmation dialog

Parameter completionHandler: A completion block that is called after the request finishes. errorThe error will be non-nil if the request does not result in a window being shared. The error will be NSUserCancelledError if there is no ScreenCaptureKit session, or if the user rejects the offer to share. If sharing fails for some other reason, the error will provide the details.

pub fn windowTitlebarLayoutDirection(&self) -> NSUserInterfaceLayoutDirection

Retrieve the layout direction of the window titlebar: this includes the standard window buttons (close/minimize/maximize buttons) and the title for this window. In general, this will return “right to left” (RTL) if the primary system language is RTL. The layout direction may be RTL even in applications that do not have a RTL language localization. This value should be utilized if an application uses titlebarAppearsTransparent and places controls underneath the titlebar.

impl NSWindow

Methods declared on superclass NSResponder.

pub unsafe fn init(this: Allocated<NSWindow>) -> Retained<NSWindow>

impl NSWindow

Methods declared on superclass NSObject.

pub unsafe fn new(mtm: MainThreadMarker) -> Retained<NSWindow>

impl NSWindow

NSEvent.

pub fn trackEventsMatchingMask_timeout_mode_handler( &self, mask: NSEventMask, timeout: f64, mode: &NSString, tracking_handler: &Block<dyn Fn(*mut NSEvent, NonNull<Bool>) + '_>, )

Tracks events matching the supplied mask with the supplied tracking handler until the tracking handler explicitly terminates tracking. Each event is removed from the event queue then passed to the tracking handler. If a matching event does not exist in the event queue, then the main thread blocks in the specified runloop mode until an event of the requested type is received or the timeout expires. If the timeout expires, the tracking handler is called with a nil event. A negative timeout is interpreted as 0. Use NSEventDurationForeverto never timeout. Tracking continues until *stop is set to YES.Calls to -nextEventMatchingMask:… are allowed inside the trackingHandler block. This method returns once tracking is terminated.

pub fn nextEventMatchingMask( &self, mask: NSEventMask, ) -> Option<Retained<NSEvent>>

pub fn nextEventMatchingMask_untilDate_inMode_dequeue( &self, mask: NSEventMask, expiration: Option<&NSDate>, mode: &NSString, deq_flag: bool, ) -> Option<Retained<NSEvent>>

pub fn discardEventsMatchingMask_beforeEvent( &self, mask: NSEventMask, last_event: Option<&NSEvent>, )

pub fn postEvent_atStart(&self, event: &NSEvent, flag: bool)

pub fn sendEvent(&self, event: &NSEvent)

pub fn currentEvent(&self) -> Option<Retained<NSEvent>>

pub fn acceptsMouseMovedEvents(&self) -> bool

pub fn setAcceptsMouseMovedEvents(&self, accepts_mouse_moved_events: bool)

Setter for acceptsMouseMovedEvents.

pub fn ignoresMouseEvents(&self) -> bool

pub fn setIgnoresMouseEvents(&self, ignores_mouse_events: bool)

Setter for ignoresMouseEvents.

pub fn mouseLocationOutsideOfEventStream(&self) -> CGPoint

impl NSWindow

NSCursorRect.

pub fn disableCursorRects(&self)

pub fn enableCursorRects(&self)

pub fn discardCursorRects(&self)

pub fn areCursorRectsEnabled(&self) -> bool

pub fn invalidateCursorRectsForView(&self, view: &NSView)

pub fn resetCursorRects(&self)

impl NSWindow

NSDrag.

pub fn beginDraggingSessionWithItems_event_source( &self, items: &NSArray<NSDraggingItem>, event: &NSEvent, source: &ProtocolObject<dyn NSDraggingSource>, ) -> Retained<NSDraggingSession>

pub unsafe fn dragImage_at_offset_event_pasteboard_source_slideBack( &self, image: &NSImage, base_location: CGPoint, initial_offset: CGSize, event: &NSEvent, pboard: &NSPasteboard, source_obj: &AnyObject, slide_flag: bool, )

👎Deprecated:
Use -[NSWindow beginDraggingSessionWithItems:event:source:] instead.

Safety

source_obj should be of the correct type.

pub fn registerForDraggedTypes(&self, new_types: &NSArray<NSString>)

pub fn unregisterDraggedTypes(&self)

impl NSWindow

NSDisplayLink.

pub unsafe fn displayLinkWithTarget_selector( &self, target: &AnyObject, selector: Sel, ) -> Retained<CADisplayLink>

Safety

target should be of the correct type.

selector must be a valid selector.

impl NSWindow

NSDeprecated.

pub fn cacheImageInRect(&self, rect: CGRect)

👎Deprecated:
This method shouldn’t be used as it doesn’t work in all drawing situations; instead, a subview should be used that implements the desired drawing behavior

pub fn restoreCachedImage(&self)

👎Deprecated:
This method shouldn’t be used as it doesn’t work in all drawing situations; instead, a subview should be used that implements the desired drawing behavior

pub fn discardCachedImage(&self)

👎Deprecated:
This method shouldn’t be used as it doesn’t work in all drawing situations; instead, a subview should be used that implements the desired drawing behavior

pub fn menuChanged(menu: &NSMenu)

👎Deprecated:
This method does not do anything and should not be called.

pub fn gState(&self) -> isize

👎Deprecated:
This method is unused and should not be called.

pub fn convertBaseToScreen(&self, point: CGPoint) -> CGPoint

👎Deprecated:
Use -convertRectToScreen: or -convertPointToScreen: instead

pub fn convertScreenToBase(&self, point: CGPoint) -> CGPoint

👎Deprecated:
Use -convertRectFromScreen or -convertPointFromScreen: instead

pub fn userSpaceScaleFactor(&self) -> f64

👎Deprecated:
Use -convertRectToBacking: and -backingScaleFactor instead

pub fn useOptimizedDrawing(&self, flag: bool)

👎Deprecated:
This method does not do anything and should not be called.

pub fn canStoreColor(&self) -> bool

👎Deprecated:
This method does not do anything and should not be called.

pub fn disableFlushWindow(&self)

👎Deprecated:
Use +[NSAnimationContext runAnimationGroup:completionHandler:] to perform atomic updates across runloop invocations.

pub fn enableFlushWindow(&self)

👎Deprecated:
Use +[NSAnimationContext runAnimationGroup:completionHandler:] to perform atomic updates across runloop invocations.

pub fn isFlushWindowDisabled(&self) -> bool

👎Deprecated:
Use +[NSAnimationContext runAnimationGroup:completionHandler:] to perform atomic updates across runloop invocations.

pub fn flushWindow(&self)

👎Deprecated:
Allow AppKit’s automatic deferred display mechanism to take care of flushing any graphics contexts as needed.

pub fn flushWindowIfNeeded(&self)

👎Deprecated:
Allow AppKit’s automatic deferred display mechanism to take care of flushing any graphics contexts as needed.

pub fn isAutodisplay(&self) -> bool

👎Deprecated:
Use +[NSAnimationContext runAnimationGroup:completionHandler:] to temporarily prevent AppKit’s automatic deferred display mechanism from drawing.

pub fn setAutodisplay(&self, autodisplay: bool)

👎Deprecated:
Use +[NSAnimationContext runAnimationGroup:completionHandler:] to temporarily prevent AppKit’s automatic deferred display mechanism from drawing.

Setter for isAutodisplay.

pub fn graphicsContext(&self) -> Option<Retained<NSGraphicsContext>>

👎Deprecated:
Add instances of NSView to display content in a window.

pub fn isOneShot(&self) -> bool

👎Deprecated:
This property does not do anything and should not be used

pub fn setOneShot(&self, one_shot: bool)

👎Deprecated:
This property does not do anything and should not be used

Setter for isOneShot.

pub fn preferredBackingLocation(&self) -> NSWindowBackingLocation

👎Deprecated:
This property does not do anything and should not be used

pub fn setPreferredBackingLocation( &self, preferred_backing_location: NSWindowBackingLocation, )

👎Deprecated:
This property does not do anything and should not be used

Setter for preferredBackingLocation.

pub fn backingLocation(&self) -> NSWindowBackingLocation

👎Deprecated:
This property does not do anything and should not be used

pub fn showsResizeIndicator(&self) -> bool

👎Deprecated:
This property does not do anything and should not be used.

pub fn setShowsResizeIndicator(&self, shows_resize_indicator: bool)

👎Deprecated:
This property does not do anything and should not be used.

Setter for showsResizeIndicator.

pub unsafe fn initWithWindowRef( this: Allocated<NSWindow>, window_ref: NonNull<c_void>, ) -> Option<Retained<NSWindow>>

👎Deprecated:
This method should not be used.

Safety

window_ref must be a valid pointer.

pub fn windowRef(&self) -> NonNull<c_void>

👎Deprecated:
This method should not be used.

pub fn disableScreenUpdatesUntilFlush(&self)

👎Deprecated:
This method does not do anything and should not be called.

impl NSWindow

NSUserInterfaceRestoration.

pub fn isRestorable(&self) -> bool

pub fn setRestorable(&self, restorable: bool)

Setter for isRestorable.

pub unsafe fn restorationClass(&self) -> Option<&'static AnyClass>

Safety

This is not retained internally, you must ensure the object is still alive.

pub unsafe fn setRestorationClass(&self, restoration_class: Option<&AnyClass>)

Setter for restorationClass.

Safety

restoration_class must implement NSWindowRestoration.

This is unretained, you must ensure the object is kept alive while in use.

pub fn disableSnapshotRestoration(&self)

pub fn enableSnapshotRestoration(&self)

impl NSWindow

NSScripting.

pub fn hasCloseBox(&self) -> bool

pub fn hasTitleBar(&self) -> bool

pub fn isFloatingPanel(&self) -> bool

pub fn isMiniaturizable(&self) -> bool

pub fn isModalPanel(&self) -> bool

pub fn isResizable(&self) -> bool

pub fn isZoomable(&self) -> bool

pub fn orderedIndex(&self) -> isize

pub fn setOrderedIndex(&self, ordered_index: isize)

Setter for orderedIndex.

pub fn setIsMiniaturized(&self, flag: bool)

pub fn setIsVisible(&self, flag: bool)

pub fn setIsZoomed(&self, flag: bool)

pub fn handleCloseScriptCommand( &self, command: &NSCloseCommand, ) -> Option<Retained<AnyObject>>

pub fn handlePrintScriptCommand( &self, command: &NSScriptCommand, ) -> Option<Retained<AnyObject>>

pub fn handleSaveScriptCommand( &self, command: &NSScriptCommand, ) -> Option<Retained<AnyObject>>