tauri_nspanel

Type Alias NSPoint

pub type NSPoint = CGPoint;

A point in a Cartesian coordinate system.

This is a convenience alias for CGPoint. For ease of use, it is available on all platforms, though in practice it is only useful on macOS.

See Apple’s documentation.

Aliased Type

struct NSPoint {
    pub x: f64,
    pub y: f64,
}

Fields
x: f64

The x-coordinate of the point.

y: f64

The y-coordinate of the point.